Added some exception handling, added new exception, and refactored some Amenity classes.

+++ findmyparadise/src/main/java/com/tp/wrc/findmyparadise/services/AmenityServiceImpl.java
@@ -1,5 +1,7 @@
 package com.tp.wrc.findmyparadise.services;
 
+import com.tp.wrc.findmyparadise.exceptions.InvalidAmenityIdException;
+import com.tp.wrc.findmyparadise.exceptions.NullAmenityIdException;
 import com.tp.wrc.findmyparadise.models.Amenity;
 import com.tp.wrc.findmyparadise.models.Event;
 import com.tp.wrc.findmyparadise.repositories.AmenityRepository;
@@ -17,7 +19,7 @@ public class AmenityServiceImpl implements AmenityService {
 
     @Override
     public Amenity create(Amenity amenity) {
-        return null;
+        return amenityRepository.saveAndFlush(amenity);
     }
 
     @Override
@@ -27,7 +29,10 @@ public class AmenityServiceImpl implements AmenityService {
     }
 
     @Override
-    public Amenity getAmenityById(Integer id) {
+    public Amenity findAmenityById(Integer id) throws NullAmenityIdException, InvalidAmenityIdException {
+        if(id == null)
+            throw new NullAmenityIdException("Amenity ID is null.");
+
         Amenity amenity = null;
         Optional<Amenity> opt = amenityRepository.findById(id);
         if (opt.isPresent()) {
@@ -35,18 +40,24 @@ public class AmenityServiceImpl implements AmenityService {
 
             return amenity;
         }
-        return null;
+        else {
+            throw new InvalidAmenityIdException("Amenity does not exist.");
+        }
+
     }
 
     @Override
-    public Amenity getAmenityByCategory(String category) {
-        Amenity amenity = amenityRepository.findByCategory(category);
+    public List<Amenity> findAmenityByCategory(String category) {
+        List<Amenity> amenities = amenityRepository.findByCategory(category);
 
-        return amenity;
+        return amenities;
     }
 
     @Override
-    public Amenity update(Amenity amenity) {
+    public Amenity update(Amenity amenity) throws NullAmenityIdException, InvalidAmenityIdException{
+        if(amenity.getAmenityId() == null)
+            throw new NullAmenityIdException("Amenity ID is null.");
+
         Amenity toUpdate = amenityRepository.findById(amenity.getAmenityId()).get();
 
         if(toUpdate != null) {
@@ -56,19 +67,26 @@ public class AmenityServiceImpl implements AmenityService {
 
             return amenityRepository.saveAndFlush(toUpdate);
         }
+        else {
+            throw new InvalidAmenityIdException("Amenity does not exist.");
+        }
 
-        return null;
     }
 
     @Override
-    public boolean destroy(Integer id) {
+    public boolean destroy(Integer id) throws NullAmenityIdException, InvalidAmenityIdException{
+        if(id == null)
+            throw new NullAmenityIdException("Amenity ID is null.");
+
         Amenity amenity = amenityRepository.findById(id).get();
 
         if (amenity != null) {
             amenityRepository.delete(amenity);
             return true;
         }
+        else {
+            throw new InvalidAmenityIdException("Amenity does not exist.");
+        }
 
-        return false;
     }
 }
